+++ src/InlineInput/NumberInlineInput.tsx
@@ -30,6 +30,8 @@ interface NumberInlineInput {
     textAlignLeft?: boolean;
     onValidityChanged?: (validity: boolean) => void;
     preventDefaultInvalidStyle?: boolean;
+    minSize?: number;
+    preAllocateSize?: boolean;
 }
 
 const isInValues = (value: number, values: Values) => values.includes(value);
@@ -106,6 +108,8 @@ export default ({
     textAlignLeft,
     onValidityChanged,
     preventDefaultInvalidStyle,
+    minSize,
+    preAllocateSize = true,
 }: NumberInlineInput) => {
     useValidatedRange(range);
 
@@ -143,6 +147,25 @@ export default ({
             }}
             onValidityChanged={onValidityChanged}
             preventDefaultInvalidStyle={preventDefaultInvalidStyle}
+            minSize={
+                preAllocateSize
+                    ? minSize ??
+                      Math.max(
+                          ...(isValues(range)
+                              ? range.map(v =>
+                                    Number.isFinite(v)
+                                        ? handleInfinityToString(v)
+                                        : v.toString()
+                                )
+                              : [range.min, range.max].map(v =>
+                                    Number.isFinite(v)
+                                        ? v.toFixed(range.decimals)
+                                        : handleInfinityToString(v)
+                                )
+                          ).map(v => v.length)
+                      )
+                    : undefined
+            }
         />
     );
 };